less hacky
