+++ resources/resources/software/aws/toolkits/resources/MessagesBundle.properties
@@ -424,6 +424,7 @@ delete_resource.delete_failed=Failed to delete {0} ''{1}''
 delete_resource.deleted=Deleted {0} ''{1}''
 delete_resource.message=Are you sure you want to delete {0} ''{1}''?\n\nType "delete me" below to confirm.
 delete_resource.title=Delete {0} {1}
+dockerfile.building=Building Dockerfile: {0}
 dockerfile.label=Dockerfile:
 dynamodb.viewer.open.failed=Failed to open DynamoDB table
 dynamodb.viewer.open.failed.with_error=Failed to open DynamoDB table: {0}
